services/mqtt-farmos-logger/requirements.txt
paho-mqtt>=1.6,<2.0
pyyaml>=6.0
jsonschema>=4.21
pydantic>=2.6
requests>=2.31